+++ analysis/src/Completions.ml
@@ -9,6 +9,14 @@ let getCompletions ~debug ~path ~pos ~currentFile ~forHover =
     with
     | None -> None
     | Some (completable, scope) -> (
+      (* uncomment when debugging *)
+      if false then (
+        Printf.printf "\nScope from frontend:\n";
+        List.iter
+          (fun item ->
+            Printf.printf "%s\n" (SharedTypes.ScopeTypes.item_to_string item))
+          scope;
+        print_newline ());
       (* Only perform expensive ast operations if there are completables *)
       match Cmt.loadFullCmtFromPath ~path with
       | None -> None